+++ unit_tests/E002_Normal_Exception_traceback.py
@@ -1,4 +1,0 @@
-#
-print("hello world")
-print(1/0)
-print("goodbye world")