+++ unit_tests/E003_exit_enhanced.py
@@ -1,5 +1,0 @@
-#! /usr/bin/env python -m pydblclick
-
-print("hello")
-exit(1)
-print("your won't see this")