+++ unit_tests/E003_quit_enhanced.py
@@ -1,5 +1,0 @@
-#! /usr/bin/env python -m pydblclick
-
-print("hello")
-quit(1)
-print("your won't see this")