+++ unit_tests/E003_systemexit_enhanced.py
@@ -1,5 +1,0 @@
-#! /usr/bin/env python -m pydblclick
-
-print("hello")
-raise SystemExit
-print("your won't see this")