+++ unit_tests/E004_reading_dunder_file_enhanced.py
@@ -1,4 +1,0 @@
-#!/usr/bin/env python -m pydblclick
-import pprint as pp
-print("__file__=" + __file__)
-pp.pprint(globals())