+++ unit_tests/E004_reading_dunder_file_normal.py
@@ -1,6 +1,0 @@
-#
-import pprint as pp
-print("__file__=" + __file__)
-pp.pprint(globals())
-
-input("pause")